Added probability coloring of hypnogram.
`Annotations` hash the maximum probability values in `max_prob`, so that
these can be used by `Hypnogram` to assign a color.  Unfortunately
plotly doesn't support line segment coloring, which is why squares are
plotted on top of the line.  The color is between p=0.5 (red) and p=1.0
(green) and is clamped beyond.